Extracting common component for code reuse
