+++ frontend/src/features/permits/components/permit-list/PermitRowOptions.tsx
@@ -1,9 +1,4 @@
-import * as React from "react";
-import IconButton from "@mui/material/IconButton";
-import Menu from "@mui/material/Menu";
-import MenuItem from "@mui/material/MenuItem";
-import MoreVertIcon from "@mui/icons-material/MoreVert";
-import { useCallback } from "react";
+import { OnRouteBCTableRowActions } from "../../../../common/components/table/OnRouteBCTableRowActions";
 import { downloadReceiptPdf } from "../../apiManager/permitsAPI";
 import { openBlobInNewTab } from "../../helpers/openPdfInNewTab";
 
@@ -26,15 +21,10 @@ export const PermitRowOptions = ({
   isExpired: boolean;
   permitId: number;
 }) => {
-  const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null);
-  const open = Boolean(anchorEl);
-  const handleClick = useCallback((event: React.MouseEvent<HTMLElement>) => {
-    setAnchorEl(event.currentTarget);
-  }, []);
-  const handleClose = (_event: React.MouseEvent<HTMLElement>) => {
-    setAnchorEl(null);
-  };
-
+  /**
+   * Opens the receipt pdf in a new tab.
+   * @param permitId The permit id.
+   */
   const viewReceiptPdf = async (permitId: string) => {
     if (permitId) {
       try {
@@ -48,48 +38,22 @@ export const PermitRowOptions = ({
     }
   };
 
-  const onClickOption = (event: React.MouseEvent<HTMLElement>) => {
-    const selectedOption = event.currentTarget.outerText as string;
-
+  /**
+   * Action handler upon a select event.
+   * @param selectedOption The option that was selected.
+   */
+  const onSelectOptionCallback = (selectedOption: string) => {
     if (selectedOption === "View Receipt") {
       viewReceiptPdf(permitId.toString());
     }
-    setAnchorEl(null);
   };
 
   return (
     <>
-      <IconButton
-        aria-label="more"
-        id="long-button"
-        aria-controls={open ? "long-menu" : undefined}
-        aria-expanded={open ? "true" : undefined}
-        aria-haspopup="true"
-        onClick={handleClick}
-      >
-        <MoreVertIcon />
-      </IconButton>
-      <Menu
-        id="long-menu"
-        MenuListProps={{
-          "aria-labelledby": "long-button",
-        }}
-        anchorEl={anchorEl}
-        open={open}
-        onClose={handleClose}
-        PaperProps={{
-          style: {
-            maxHeight: ITEM_HEIGHT * 4.5,
-            width: "15ch",
-          },
-        }}
-      >
-        {getOptions(isExpired).map((option) => (
-          <MenuItem key={option} onClick={onClickOption} option-value={option}>
-            {option}
-          </MenuItem>
-        ))}
-      </Menu>
+      <OnRouteBCTableRowActions
+        onSelectOption={onSelectOptionCallback}
+        options={getOptions(isExpired)}
+      />
     </>
   );
 };
